+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Provide access to the customer automation SA for managing workload cluster client certificates.
+
 ## [0.17.0] - 2021-11-04
 
 ### Added